+++ bench/source/reduce/omp.cpp
@@ -9,7 +9,7 @@
 
 namespace {
 
-auto reduce(std::span<float> x, std::size_t grain_size) -> float {
+auto reduce(std::span<double> x, std::size_t grain_size) -> double {
   //
   if (x.size() <= grain_size) {
     return std::reduce(x.begin(), x.end());
@@ -18,7 +18,7 @@ auto reduce(std::span<float> x, std::size_t grain_size) -> float {
   auto h = x.size() / 2;
   auto t = x.size() - h;
 
-  float a, b;
+  double a, b;
 
 #pragma omp task untied shared(a, x, grain_size, h) default(none)
   a = reduce(x.first(h), grain_size);
@@ -37,7 +37,7 @@ void reduce_omp(benchmark::State &state) {
 #pragma omp parallel num_threads(n)
 #pragma omp single
   {
-    std::vector<float> data;
+    std::vector<double> data;
 
 #pragma omp task shared(data) // No untied
     data = to_sum();
@@ -46,7 +46,7 @@ void reduce_omp(benchmark::State &state) {
 
     auto grain_size = data.size() / (n * 10);
 
-    volatile float output;
+    volatile double output;
 
     for (auto _ : state) {
       output = reduce(data, grain_size);
